requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
langchain-anthropic==0.3.7
python-dotenv==1.0.1
langchain-google-genai==2.0.10
langchain-openai==0.3.6
langchain-chroma==0.2.2
langchain-pinecone==0.2.3
pymongo==4.11.1
langchain-community==0.3.18
BeautifulSoup4==4.13.5
jq==1.8.0
streamlit==1.36.0
selenium==4.35.0
requests==2.32.5
webdriver_manager==4.0.2
unstructured==0.16.23
pdfminer.six==20240706
pi_heif==0.21.0
unstructured-inference==0.8.7
unstructured.pytesseract==0.3.13
pdf2image==1.17.0
wikipedia==1.4.0
youtube-transcript-api==0.6.3
